XXX make exception an owned pointer
Exception was 88 bytes in size, so every function that returned
Result<T, Exception> was copying a big block around. For the cost of one
allocation per exception, we can reduce this to a pointer copy. For the
common case of Result<Value, Exception>, this makes both sides the same
size.